[FIXES #12076] Move the new users API under an api module
