+++ geonode/urls.py
@@ -42,7 +42,7 @@ from geonode import geoserver
 from geonode.utils import check_ogc_backend
 from geonode.base import register_url_event
 from geonode.messaging.urls import urlpatterns as msg_urls
-from .people.views import CustomSignupView
+from .people.api.views import CustomSignupView
 from oauth2_provider.urls import app_name as oauth2_app_name, base_urlpatterns, oidc_urlpatterns
 
 admin.autodiscover()
@@ -93,7 +93,7 @@ urlpatterns += [
     re_path(r"^account/signup/", CustomSignupView.as_view(), name="account_signup"),
     re_path(r"^account/", include("allauth.urls")),
     re_path(r"^invitations/", include("geonode.invitations.urls", namespace="geonode.invitations")),
-    re_path(r"^people/", include("geonode.people.urls")),
+    re_path(r"^people/", include("geonode.people.api.urls")),
     re_path(r"^avatar/", include("avatar.urls")),
     re_path(r"^activity/", include("actstream.urls")),
     re_path(r"^announcements/", include("announcements.urls")),
